debug progress

+++ ann.h
@@ -164,13 +164,13 @@ struct Network
 int ann_add_layer(PNetwork pnet, int node_count, Layer_type layer_type, Activation_type activation_type);
 PNetwork ann_make_network(Optimizer_type opt, Loss_type loss_type);
 void ann_free_network(PNetwork pnet);
-int ann_load_csv(const char *filename, int has_header, real **data, size_t *rows, size_t *stride);
+int ann_load_csv(const char *filename, int has_header, real **data, int *rows, int *stride);
 PNetwork ann_load_network(const char *filename);
 int ann_save_network(PNetwork pnet, const char *filename);
 int ann_save_network_binary(PNetwork pnet, const char *filename);
 
 // training/evaluating
-real ann_train_network(PNetwork pnet, PTensor inputs, PTensor outputs, size_t rows);
+real ann_train_network(PNetwork pnet, PTensor inputs, PTensor outputs, int rows);
 void ann_set_convergence(PNetwork pnet, real limit);
 int ann_predict(PNetwork pnet, real *inputs, real *outputs);
 int ann_class_prediction(real *outputs, int classes);
